put sentinel back
